feat(74qFSWXi): make authorities work.
There were problems with enum field.
Since Postgres is ORM it waits only for
objects, but Hibernate doesn't know that
and send VARCHAR to enum field.

Via new dependency (in build.gradle.kts)
I tell to Hibernate how it should map
this fields to Postgres table.

I was solving this issue for 4 days.